requirements.txt
alembic==1.14.1
aniso8601==10.0.0
attrs==24.3.0
bcrypt==4.2.1
blinker==1.8.2
Brotli==1.1.0
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2024.8.30
cffi==1.17.1
click==8.1.8
colorama==0.4.6
contourpy==1.1.1
cssselect2==0.7.0
cycler==0.12.1
defusedxml==0.7.1
distlib==0.3.9
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
Faker==35.2.0
filelock==3.16.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.54.1
greenlet==3.1.1
gunicorn==23.0.0
html5lib==1.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
Mako==1.3.8
MarkupSafe==2.1.5
matplotlib==3.7.5
numpy==1.24.4
packaging==24.1
pandas==2.0.3
pillow==10.4.0
pipenv==2024.2.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
playsound==1.3.0
pluggy==1.5.0
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.3.0
pygame==2.6.1
PyJWT==2.9.0
pyparsing==3.1.4
pyphen==0.16.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
rpds-py==0.20.1
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
six==1.16.0
SQLAlchemy==2.0.37
SQLAlchemy-serializer==1.4.12
threadpoolctl==3.5.0
tinycss2==1.4.0
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2024.2
virtualenv==20.27.0
WeasyPrint==52.5
webencodings==0.5.1
Werkzeug==3.0.6
WTForms==3.1.2
zipp==3.20.2
zopfli==0.2.3.post1
